backend/requirements.txt
ocp-tessellate
cadquery
ocp-vscode
openai
flask
python-dotenv
flask-cors
schemdraw
git+https://github.com/meadiode/cq_gears.git@main
git+https://github.com/OpenOrion/parafoil.git#egg=parafoil